inference/requirements.txt
openai>=1.48.0
Jinja2>=3.1.4
PyYAML>=6.0.1
tqdm>=4.66.5
python-dotenv>=1.0.1

